requirements.txt
python-telegram-bot==20.7
requests==2.32.3
apscheduler==3.10.4
